compoent split
